+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/peers/Eth2PeerSelectionStrategyTest.java
@@ -20,7 +20,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static tech.pegasys.teku.util.config.Constants.ATTESTATION_SUBNET_COUNT;
 
-import com.google.common.primitives.Ints;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.util.Comparator;
@@ -309,6 +308,6 @@ class Eth2PeerSelectionStrategyTest {
         peerId,
         new InetSocketAddress(InetAddress.getLoopbackAddress(), peerId.trimLeadingZeros().toInt()),
         ENR_FORK_ID,
-        new Bitvector(Ints.asList(subnetIds), ATTESTATION_SUBNET_COUNT));
+        new Bitvector(ATTESTATION_SUBNET_COUNT, subnetIds));
   }
 }