+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - Restored the state cache size to 160 to improve performance during sync.
 - Fixed help text for `--validators-graffiti-file` to refer to `--validators-graffiti` as the fallback not `--graffiti`.
 
+### Additions and Improvements
+- `--validators-external-signer-public-keys` now accepts a URL to load public keys from.
+
 ## 21.1.0
 
 ### Additions and Improvements